Revert "	modified:   Assembly/AssemblyAlgorithms.cpp"
This reverts commit e442ae09809e8161f5ea4abf4935ca8aae37e908.

+++ ParseAligns/abyss-fixmate.cc
@@ -210,7 +210,7 @@ static void readAlignmentsFile(string path, Alignments* pMap)
 static string percent(size_t x, size_t n)
 {
 	ostringstream ss;
-	ss << setw((int)log10((double)n) + 1) << x;
+	ss << setw((int)log10(n) + 1) << x;
 	if (x > 0)
 		ss << "  " << setprecision(3) << (float)100*x/n << '%';
 	return ss.str();
